@@ -11,15 +11,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.13",
-    "@angular/common": "^7.2.14",
-    "@angular/compiler": "^7.2.13",
-    "@angular/core": "^7.2.13",
-    "@angular/forms": "^7.2.13",
+    "@angular/animations": "^11.0.5",
+    "@angular/common": "^11.0.5",
+    "@angular/compiler": "^11.0.5",
+    "@angular/core": "^11.0.5",
+    "@angular/fire": "^6.1.4",
+    "@angular/forms": "^11.0.5",
     "@angular/http": "^7.2.13",
-    "@angular/platform-browser": "^7.2.13",
-    "@angular/platform-browser-dynamic": "^7.2.13",
-    "@angular/router": "^7.2.13",
+    "@angular/platform-browser": "^11.0.5",
+    "@angular/platform-browser-dynamic": "^11.0.5",
+    "@angular/router": "^11.0.5",
     "bootstrap": "^4.3.1",
     "braces": "^3.0.2",
     "core-js": "^2.5.4",
@@ -33,15 +34,16 @@
     "ngx-spinner": "^7.1.4",
     "ngx-toastr": "^10.0.2",
     "popper.js": "^1.15.0",
-    "rxjs": "^6.4.0",
+    "rxjs": "^6.6.3",
     "ws": "^7.4.1",
-    "zone.js": "^0.8.29"
+    "zone.js": "^0.10.3",
+    "firebase": "^7.0 || ^8.0"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.1100.5",
     "@angular/cli": "^11.0.5",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/language-service": "^7.2.13",
+    "@angular/compiler-cli": "^11.0.5",
+    "@angular/language-service": "^11.0.5",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
@@ -56,6 +58,12 @@
     "protractor": "^5.4.1",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~3.2.4"
+    "typescript": "~4.0.5",
+    "@angular-devkit/architect": ">= 0.900 < 0.1200",
+    "firebase-tools": "^8.0.0",
+    "fuzzy": "^0.1.3",
+    "inquirer": "^6.2.2",
+    "inquirer-autocomplete-prompt": "^1.0.1",
+    "open": "^7.0.3"
   }
 }
