@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { SerializeParamComponent } from './serialize-param.component';
 
@@ -6,7 +6,7 @@ describe('SerializeParamComponent', () => {
   let component: SerializeParamComponent;
   let fixture: ComponentFixture<SerializeParamComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ SerializeParamComponent ]
     })
